change __call__ return to link with zellij

+++ optimization/model_evaluation/model_eval.py
@@ -43,8 +43,9 @@ class ModelEvaluator:
         self.model_name = self.models[self.model_id]
         self.exp_path = self.experiment["experiment_path"]
         self.hp_key = list(self.hyperparameters.keys())
-        with open(self.experiment["historic_file"], 'w') as f:
-            pass
+        if not Path(self.experiment["historic_file"]).is_file():         
+            with open(self.experiment["historic_file"], 'w') as f:
+                pass
         os.environ["PYTORCH_CUDA_ALLOC_CONF"] = "garbage_collection_threshold:0.3"
 
     def load_config(self,config_file):    
@@ -273,7 +274,7 @@ class ModelEvaluator:
         float
             Result of the evaluation for the first task.
         """
-        return self.evaluate(x)
+        return {"obj":self.evaluate(x)}
 
 if __name__ == "__main__":
     evaluator = ModelEvaluator()
